+++ src/main/java/com/kgu/studywithme/global/annotation/validation/ValidImageExtension.java
@@ -1,4 +1,4 @@
-package com.kgu.studywithme.global.annotation;
+package com.kgu.studywithme.global.annotation.validation;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
