+++ src/main/java/com/kgu/studywithme/upload/controller/dto/request/ImageUploadRequest.java
@@ -1,6 +1,6 @@
 package com.kgu.studywithme.upload.controller.dto.request;
 
-import com.kgu.studywithme.global.annotation.ValidImageExtension;
+import com.kgu.studywithme.global.annotation.validation.ValidImageExtension;
 import org.springframework.web.multipart.MultipartFile;
 
 public record ImageUploadRequest(
