+++ src/test/java/com/kgu/studywithme/global/annotation/ValidHashtagCountValidatorTest.java
@@ -1,5 +1,7 @@
 package com.kgu.studywithme.global.annotation;
 
+import com.kgu.studywithme.global.annotation.validation.ValidHashtagCount;
+import com.kgu.studywithme.global.annotation.validation.ValidHashtagCountValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
