+++ src/test/java/com/kgu/studywithme/global/annotation/ValidImageExtensionValidatorTest.java
@@ -1,5 +1,6 @@
 package com.kgu.studywithme.global.annotation;
 
+import com.kgu.studywithme.global.annotation.validation.ValidImageExtensionValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
